+++ src/bp/core/database/tables/server-wide/index.ts
@@ -3,4 +3,4 @@ export * from './metadata'
 export * from './data_retention'
 export * from './workspace_users'
 export * from './workspace_invite_codes'
-export * from './telemetry_payloads'
+export * from './telemetry'
