+++ src/bp/core/database/tables/server-wide/telemetry.ts
